+++ web/usuaris/src/helpers/fetchers.ts
@@ -8,3 +8,35 @@ export const getUsers = async () => {
   const data = await response.json()
   return data
 }
+
+export const loginUserFetcher = async (user: object) => {
+  const errorMessage: { [key: number]: string } = {
+    401: 'Credenciales incorrectas.',
+    403: 'Sólo accesible a usuarios administradores activos.',
+  }
+
+  const response = await fetch(urls.logIn, {
+    method: 'POST',
+    body: JSON.stringify(user),
+    headers: { 'Content-Type': 'application/json' },
+  })
+
+  if (
+    !response.ok &&
+    Object.hasOwnProperty.call(errorMessage, response.status)
+  ) {
+    throw new Error(errorMessage[response.status])
+  }
+
+  return response.status === 204 ? null : response.json()
+}
+
+export const userInfoFetcher = async () => {
+  const response = await fetch(urls.getMe)
+
+  if (!response.ok) {
+    throw new Error(`Error fetching user info`)
+  }
+
+  return response.json()
+}