+++ web/usuaris/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useGetUsers } from './useGetUsers'
+export { useLogin } from './useLogin'