+++ web/usuaris/src/pages/Home.tsx
@@ -3,16 +3,18 @@ import styled from 'styled-components'
 import { FlexBox, colors, dimensions } from '@itacademy/ui'
 import {
   FiltersWidget,
+  Login,
   Navbar,
   SideMenu,
   UsersTable,
 } from '../components/organisms'
+import { useAuth } from '../context/AuthProvider'
 
 const Container = styled(FlexBox)`
   width: 100%;
   height: 100vh;
   background-color: ${colors.gray.gray5};
-  padding: ${dimensions.spacing.none} ${dimensions.spacing.xxxl}
+  padding: ${dimensions.spacing.xxs} ${dimensions.spacing.xxxl}
     ${dimensions.spacing.xl} ${dimensions.spacing.sm};
 `
 
@@ -30,20 +32,47 @@ const MainDiv = styled(FlexBox)`
   color: ${colors.gray.gray3};
 `
 
-export const Home: FC = () => (
-  <Container direction="row" align="center">
-    <SideMenu />
-    <ContainerMain justify="flex-start">
-      <Navbar />
-      <MainDiv
-        as="main"
-        justify="flex-start"
+const LoginContainer = styled(FlexBox)`
+  width: 50%;
+  height: auto;
+  background-color: ${colors.white};
+  padding: ${dimensions.spacing.lg} 6rem;
+  border-radius: ${dimensions.borderRadius.sm};
+`
+
+export const Home: FC = () => {
+  const { user } = useAuth()
+
+  if (!user)
+    return (
+      <Container
+        direction="column"
         align="center"
-        gap={dimensions.spacing.xl}
+        justify="flex-start"
+        gap={dimensions.spacing.sm}
       >
-        <FiltersWidget />
-        <UsersTable />
-      </MainDiv>
-    </ContainerMain>
-  </Container>
-)
+        <Navbar />
+        <LoginContainer>
+          <Login />
+        </LoginContainer>
+      </Container>
+    )
+
+  return (
+    <Container direction="row" align="center">
+      <SideMenu />
+      <ContainerMain justify="flex-start">
+        <Navbar />
+        <MainDiv
+          as="main"
+          justify="flex-start"
+          align="center"
+          gap={dimensions.spacing.xxl}
+        >
+          <FiltersWidget />
+          <UsersTable />
+        </MainDiv>
+      </ContainerMain>
+    </Container>
+  )
+}