+++ services/sso/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.17.0] - 2024-04-09
+
+### Added
+
+- GET `dashboard/users/me` endpoint.
+
+## [1.16.2] - 2024-04-08
+
+### Changed
+
+- `authenticate` middleware refactored in one function.
+
 ## [1.16.1] - 2024-03-20
 
 ### Changed